#3 [fix] 리뷰 점수 타입 변경 및 이미지 업로드 로직 순서 수정

+++ src/main/java/com/dessert/gallery/dto/store/list/ReviewListDto.java
@@ -18,7 +18,7 @@ public class ReviewListDto {
     private String content;
 
     @ApiModelProperty(value = "리뷰 점수")
-    private int score;
+    private Double score;
 
     @ApiModelProperty(value = "리뷰 작성 일자")
     private LocalDateTime createDate;
